[ui, refactor] Move dashboard page under app/dashboard
